Public Documentation

Documentation for EDM4hep.jl public interface.

Index - Types

EDM4hep.EDCollection

EDM4hep.edm4hep!CaloHitContribution

EDM4hep.edm4hep!CalorimeterHit

EDM4hep.edm4hep!Cluster

EDM4hep.edm4hep!CovMatrix2f

EDM4hep.edm4hep!CovMatrix3f

EDM4hep.edm4hep!CovMatrix4f

EDM4hep.edm4hep!CovMatrix6f

EDM4hep.edm4hep!EventHeader

EDM4hep.edm4hep!GeneratorEventParameters

EDM4hep.edm4hep!MCParticle

EDM4hep.edm4hep!ParticleID

EDM4hep.edm4hep!Quantity

EDM4hep.edm4hep!RawCalorimeterHit

EDM4hep.edm4hep!RawTimeSeries

EDM4hep.edm4hep!RecDqdx

EDM4hep.edm4hep!ReconstructedParticle

EDM4hep.edm4hep!SenseWireHit

EDM4hep.edm4hep!SimCalorimeterHit

EDM4hep.edm4hep!SimTrackerHit

EDM4hep.edm4hep!TimeSeries

EDM4hep.edm4hep!Track

EDM4hep.edm4hep!TrackState

EDM4hep.edm4hep!TrackerHit

EDM4hep.edm4hep!TrackerHit3D

EDM4hep.edm4hep!TrackerHitPlane

EDM4hep.edm4hep!Vector2f

EDM4hep.edm4hep!Vector3d

EDM4hep.edm4hep!Vector3f

EDM4hep.edm4hep!Vector4f

EDM4hep.edm4hep!Vertex

EDM4hep.RootIO.Reader

EDM4hep.Histograms.H1D

EDM4hep.Histograms.H2D

EDM4hep.Histograms.H3D

Index - Functions

EDM4hep.emptyEDStore

EDM4hep.getEDCollection

EDM4hep.hasEDCollection

EDM4hep.initEDCollection

EDM4hep.RootIO.create_getter

EDM4hep.RootIO.get

EDM4hep.RootIO.get

EDM4hep.Analysis.do_analysis!

Modules

EDM4hep.EDM4hep — Module

Main module for EDM4hep.jl – Key4hep Event Data Model for Julia.

All data model types are exported from this module for public use

Exports

EDM4hep.RootIO — Module

ROOT I/O module for EDM4hep.jl

It supports both formats: TTree and RNTuple

EDM4hep.Analysis — Module

Analysis module for EDM4hep.jl

A number of data structures and functions to support analysis of EDM4hep data in multithreaded mode.

Types

This is the list of all types defined for EDM4hep using the PODIO yaml file.

EDM4hep.EDCollection — Type

EDCollection{ED} where ED <: POD

EDM4hep.edm4hep!CaloHitContribution — Type

Monte Carlo contribution to SimCalorimeterHit

Author: EDM4hep authors

Fields

PDG::Int32: PDG code of the shower particle that caused this contribution

energy::Float32: energy of this contribution

time::Float32: time of this contribution

stepPosition::edm4hep!Vector3f: position of this energy deposition (step)

stepLength::Float32: Geant4 step length for this contribution

Relations

particle::edm4hep!MCParticle: MCParticle responsible for this contribution to the hit. Only particles that are kept in the MCParticle record will appear here. Hence, this will point to the first mother appearing in the record.

EDM4hep.edm4hep!CalorimeterHit — Type

Calorimeter hit

Author: EDM4hep authors

Fields

cellID::UInt64: detector specific (geometrical) cell id

energy::Float32: energy of the hit

energyError::Float32: error of the hit energy

time::Float32: time of the hit

position::edm4hep!Vector3f: position of the hit in world coordinates

type::Int32: type of hit

EDM4hep.edm4hep!Cluster — Type

Calorimeter Hit Cluster

Author: EDM4hep authors

Fields

type::Int32: flagword that defines the type of cluster

energy::Float32: energy of the cluster

energyError::Float32: error on the energy

position::edm4hep!Vector3f: position of the cluster

positionError::edm4hep!CovMatrix3f: covariance matrix of the position

iTheta::Float32: Polar angle of the cluster's intrinsic direction (used e.g. for vertexing). Not to be confused with the cluster position seen from IP

iPhi::Float32: Azimuthal angle of the cluster's intrinsic direction (used e.g. for vertexing). Not to be confused with the cluster position seen from IP

directionError::edm4hep!Vector3f: covariance matrix of the direction

shapeParameters::PVector{Float32}: shape parameters. The corresponding names of the shape parameters should be stored in the collection named by edm4hep::labels::ShapeParameterNames in the file-level metadata, as a vector of strings in the same order as the parameters.

subdetectorEnergies::PVector{Float32}: energy observed in a particular subdetector

Relations

clusters::edm4hep!Cluster: clusters that have been combined to this cluster

hits::edm4hep!CalorimeterHit: hits that have been combined to this cluster

Methods

setShapeParameters(object::edm4hep!Cluster, v::AbstractVector{Float32}): assign a set of values to the shapeParameters vector member

setSubdetectorEnergies(object::edm4hep!Cluster, v::AbstractVector{Float32}): assign a set of values to the subdetectorEnergies vector member

pushToClusters(obj::edm4hep!Cluster, robj::edm4hep!Cluster): push related object to the clusters relation

popFromClusters(obj::edm4hep!Cluster): pop last related object from clusters relation

pushToHits(obj::edm4hep!Cluster, robj::edm4hep!CalorimeterHit): push related object to the hits relation

popFromHits(obj::edm4hep!Cluster): pop last related object from hits relation

EDM4hep.edm4hep!CovMatrix2f — Type

A generic 2 dimensional covariance matrix with values stored in lower triangular form

Fields

values::SVector{3,Float32}: the covariance matrix values

EDM4hep.edm4hep!CovMatrix3f — Type

A generic 3 dimensional covariance matrix with values stored in lower triangular form

Fields

values::SVector{6,Float32}: the covariance matrix values

EDM4hep.edm4hep!CovMatrix4f — Type

A generic 4 dimensional covariance matrix with values stored in lower triangular form

Fields

values::SVector{10,Float32}: the covariance matrix values

EDM4hep.edm4hep!CovMatrix6f — Type

A generic 6 dimensional covariance matrix with values stored in lower triangular form

Fields

values::SVector{21,Float32}: the covariance matrix values

EDM4hep.edm4hep!EventHeader — Type

Event Header. Additional parameters are assumed to go into the metadata tree.

Author: EDM4hep authors

Fields

eventNumber::UInt64: event number

runNumber::UInt32: run number

timeStamp::UInt64: time stamp

weight::Float64: event weight

weights::PVector{Float64}: event weights in case there are multiple. NOTE that weights[0] might not be the same as weight! The corresponding names of the event weights should be stored in the collection named by edm4hep::labels::EventWeightsNames in the file-level metadata.

Methods

setWeights(object::edm4hep!EventHeader, v::AbstractVector{Float64}): assign a set of values to the weights vector member

EDM4hep.edm4hep!GeneratorEventParameters — Type

Generator Event Parameters and information

Author: EDM4hep authors

Fields

sqrts::Float64: sqrt(s) - The nominal beam center of mass energy

beamsPz::SVector{2,Float64}: nominal z-momentum of the two incoming particle (beams)

partonIds::SVector{2,Int32}: PDG id of the partons undergoing the hard scatter

beamPolarizations::SVector{2,Float32}: Polarization of the incoming beam particles

crossSections::PVector{Float64}: List of cross sections

crossSectionErrors::PVector{Float64}: List of cross section errors

weights::PVector{Float64}: event weights. The corresponding names are stored using the edm4hep::labels::GeneratorWeightNames in the file level metadata.

Relations

signalVertexParticles::edm4hep!MCParticle: List of initial state MCParticles that are the source of the hard interaction

Methods

setCrossSections(object::edm4hep!GeneratorEventParameters, v::AbstractVector{Float64}): assign a set of values to the crossSections vector member

setCrossSectionErrors(object::edm4hep!GeneratorEventParameters, v::AbstractVector{Float64}): assign a set of values to the crossSectionErrors vector member

setWeights(object::edm4hep!GeneratorEventParameters, v::AbstractVector{Float64}): assign a set of values to the weights vector member

pushToSignalVertexParticles(obj::edm4hep!GeneratorEventParameters, robj::edm4hep!MCParticle): push related object to the signalVertexParticles relation

popFromSignalVertexParticles(obj::edm4hep!GeneratorEventParameters): pop last related object from signalVertexParticles relation

EDM4hep.edm4hep!MCParticle — Type

The Monte Carlo particle - based on the lcio::MCParticle.

Author: EDM4hep authors

Fields

PDG::Int32: PDG code of the particle

generatorStatus::Int32: status of the particle as defined by the generator

simulatorStatus::Int32: status of the particle from the simulation program - use BIT constants below

charge::Float32: particle charge

time::Float32: creation time of the particle in wrt. the event, e.g. for preassigned decays or decays in flight from the simulator

mass::Float64: mass of the particle

vertex::edm4hep!Vector3d: production vertex of the particle

endpoint::edm4hep!Vector3d: endpoint of the particle

momentum::edm4hep!Vector3d: particle 3-momentum at the production vertex

momentumAtEndpoint::edm4hep!Vector3d: particle 3-momentum at the endpoint

helicity::Int32: particle helicity (9 if unset)

Relations

parents::edm4hep!MCParticle: The parents of this particle

daughters::edm4hep!MCParticle: The daughters this particle

Methods

pushToParents(obj::edm4hep!MCParticle, robj::edm4hep!MCParticle): push related object to the parents relation

popFromParents(obj::edm4hep!MCParticle): pop last related object from parents relation

pushToDaughters(obj::edm4hep!MCParticle, robj::edm4hep!MCParticle): push related object to the daughters relation

popFromDaughters(obj::edm4hep!MCParticle): pop last related object from daughters relation

EDM4hep.edm4hep!ParticleID — Type

ParticleID

Author: EDM4hep authors

Fields

type::Int32: userdefined type

PDG::Int32: PDG code of this id - ( 999999 ) if unknown

algorithmType::Int32: type of the algorithm/module that created this hypothesis

likelihood::Float32: likelihood of this hypothesis - in a user defined normalization

parameters::PVector{Float32}: parameters associated with this hypothesis

Relations

particle::edm4hep!ReconstructedParticle: the particle from which this PID has been computed

Methods

setParameters(object::edm4hep!ParticleID, v::AbstractVector{Float32}): assign a set of values to the parameters vector member

EDM4hep.edm4hep!Quantity — Type

edm4hep!Quantity

Fields

type::Int32: flag identifying how to interpret the quantity

value::Float32: value of the quantity

error::Float32: error on the value of the quantity

EDM4hep.edm4hep!RawCalorimeterHit — Type

Raw calorimeter hit

Author: EDM4hep authors

Fields

cellID::UInt64: detector specific (geometrical) cell id

amplitude::Int32: amplitude of the hit in ADC counts

timeStamp::Int32: time stamp for the hit

EDM4hep.edm4hep!RawTimeSeries — Type

Raw data of a detector readout

Author: EDM4hep authors

Fields

cellID::UInt64: detector specific cell id

quality::Int32: quality flag for the hit

time::Float32: time of the hit

charge::Float32: integrated charge of the hit

interval::Float32: interval of each sampling

adcCounts::PVector{Int32}: raw data (32-bit) word at i

Methods

setAdcCounts(object::edm4hep!RawTimeSeries, v::AbstractVector{Int32}): assign a set of values to the adcCounts vector member

EDM4hep.edm4hep!RecDqdx — Type

dN/dx or dE/dx info of a Track

Author: EDM4hep authors

Fields

dQdx::edm4hep!Quantity: the reconstructed dEdx or dNdx and its error

Relations

track::edm4hep!Track: the corresponding track

EDM4hep.edm4hep!ReconstructedParticle — Type

Reconstructed Particle

Author: EDM4hep authors

Fields

PDG::Int32: PDG of the reconstructed particle.

energy::Float32: energy of the reconstructed particle. Four momentum state is not kept consistent internally

momentum::edm4hep!Vector3f: particle momentum. Four momentum state is not kept consistent internally

referencePoint::edm4hep!Vector3f: reference, i.e. where the particle has been measured

charge::Float32: charge of the reconstructed particle

mass::Float32: mass of the reconstructed particle, set independently from four vector. Four momentum state is not kept consistent internally

goodnessOfPID::Float32: overall goodness of the PID on a scale of [0;1]

covMatrix::edm4hep!CovMatrix4f: covariance matrix of the reconstructed particle 4vector

Relations

decayVertex::edm4hep!Vertex: decay vertex for the particle (if it is a composite particle)

clusters::edm4hep!Cluster: clusters that have been used for this particle

tracks::edm4hep!Track: tracks that have been used for this particle

particles::edm4hep!ReconstructedParticle: reconstructed particles that have been combined to this particle

Methods

pushToClusters(obj::edm4hep!ReconstructedParticle, robj::edm4hep!Cluster): push related object to the clusters relation

popFromClusters(obj::edm4hep!ReconstructedParticle): pop last related object from clusters relation

pushToTracks(obj::edm4hep!ReconstructedParticle, robj::edm4hep!Track): push related object to the tracks relation

popFromTracks(obj::edm4hep!ReconstructedParticle): pop last related object from tracks relation

pushToParticles(obj::edm4hep!ReconstructedParticle, robj::edm4hep!ReconstructedParticle): push related object to the particles relation

popFromParticles(obj::edm4hep!ReconstructedParticle): pop last related object from particles relation

EDM4hep.edm4hep!SenseWireHit — Type

Sense wire hit, knowing only the distance to the wire. The circle representing possible positions is parametrized with its center, radius and normal vector (given by the wire direction).

Author: EDM4hep authors

Fields

cellID::UInt64: ID of the sensor that created this hit

type::Int32: type of the raw data hit

quality::Int32: quality bit flag of the hit

time::Float32: time of the hit

eDep::Float32: energy deposited by the hit

eDepError::Float32: error on eDep

wireStereoAngle::Float32: angle between the sense wire axis and the drift chamber axis (usually the z-axis) - use it together with wireAzimuthalAngle to get the wire direction

wireAzimuthalAngle::Float32: azimuthal angle at the middle of the sense wire - use it together with wireStereoAngle to get the wire direction

position::edm4hep!Vector3d: point on the sense wire which is closest to the hit (center of the circle)

positionAlongWireError::Float64: error on the hit position along the wire direction

distanceToWire::Float32: distance between the hit and the wire (radius of the circle)

distanceToWireError::Float32: error on distanceToWire

nElectrons::PVector{UInt16}: number of electrons for each cluster (number of clusters = vector size)

Methods

setNElectrons(object::edm4hep!SenseWireHit, v::AbstractVector{UInt16}): assign a set of values to the nElectrons vector member

EDM4hep.edm4hep!SimCalorimeterHit — Type

Simulated calorimeter hit

Author: EDM4hep authors

Fields

cellID::UInt64: ID of the sensor that created this hit

energy::Float32: energy of the hit

position::edm4hep!Vector3f: position of the hit in world coordinates

Relations

contributions::edm4hep!CaloHitContribution: Monte Carlo step contributions

Methods

pushToContributions(obj::edm4hep!SimCalorimeterHit, robj::edm4hep!CaloHitContribution): push related object to the contributions relation

popFromContributions(obj::edm4hep!SimCalorimeterHit): pop last related object from contributions relation

EDM4hep.edm4hep!SimTrackerHit — Type

Simulated tracker hit

Author: EDM4hep authors

Fields

cellID::UInt64: ID of the sensor that created this hit

eDep::Float32: energy deposited in the hit

time::Float32: proper time of the hit in the lab frame

pathLength::Float32: path length of the particle in the sensitive material that resulted in this hit

quality::Int32: quality bit flag

position::edm4hep!Vector3d: the hit position

momentum::edm4hep!Vector3f: the 3-momentum of the particle at the hits position

Relations

particle::edm4hep!MCParticle: MCParticle that caused the hit

EDM4hep.edm4hep!TimeSeries — Type

Calibrated Detector Data

Author: EDM4hep authors

Fields

cellID::UInt64: cell id

time::Float32: begin time

interval::Float32: interval of each sampling

amplitude::PVector{Float32}: calibrated detector data

Methods

setAmplitude(object::edm4hep!TimeSeries, v::AbstractVector{Float32}): assign a set of values to the amplitude vector member

EDM4hep.edm4hep!Track — Type

Reconstructed track

Author: EDM4hep authors

Fields

type::Int32: flagword that defines the type of track

chi2::Float32: chi-squared of the track fit

ndf::Int32: number of degrees of freedom of the track fit

Nholes::Int32: number of holes on track

subdetectorHitNumbers::PVector{Int32}: number of hits in particular subdetectors

subdetectorHoleNumbers::PVector{Int32}: number of holes in particular subdetectors

trackStates::PVector{edm4hep!TrackState}: track states

Relations

trackerHits::edm4hep!TrackerHit: hits that have been used to create this track

tracks::edm4hep!Track: tracks (segments) that have been combined to create this track

Methods

setSubdetectorHitNumbers(object::edm4hep!Track, v::AbstractVector{Int32}): assign a set of values to the subdetectorHitNumbers vector member

setSubdetectorHoleNumbers(object::edm4hep!Track, v::AbstractVector{Int32}): assign a set of values to the subdetectorHoleNumbers vector member

setTrackStates(object::edm4hep!Track, v::AbstractVector{edm4hep!TrackState}): assign a set of values to the trackStates vector member

pushToTrackerHits(obj::edm4hep!Track, robj::edm4hep!TrackerHit): push related object to the trackerHits relation

popFromTrackerHits(obj::edm4hep!Track): pop last related object from trackerHits relation

pushToTracks(obj::edm4hep!Track, robj::edm4hep!Track): push related object to the tracks relation

popFromTracks(obj::edm4hep!Track): pop last related object from tracks relation

EDM4hep.edm4hep!TrackState — Type

Parametrized description of a particle track

Fields

location::Int32: for use with At{Other|IP|FirstHit|LastHit|Calorimeter|Vertex}|LastLocation

D0::Float32: transverse impact parameter

phi::Float32: azimuthal angle of the track at this location (i.e. not phi0)

omega::Float32: is the signed curvature of the track

Z0::Float32: longitudinal impact parameter

tanLambda::Float32: lambda is the dip angle of the track in r-z

time::Float32: time of the track at this trackstate

referencePoint::edm4hep!Vector3f: Reference point of the track parameters, e.g. the origin at the IP, or the position of the first/last hits or the entry point into the calorimeter

covMatrix::edm4hep!CovMatrix6f: covariance matrix of the track parameters.

EDM4hep.edm4hep!TrackerHit — Type

Tracker hit interface class

Author: Thomas Madlener, DESY

Fields

cellID::UInt64: ID of the sensor that created this hit

type::Int32: type of the raw data hit

quality::Int32: quality bit flag of the hit

time::Float32: time of the hit

eDep::Float32: energy deposited on the hit

eDepError::Float32: error measured on eDep

position::edm4hep!Vector3d: hit position as recorded by the sensor. The exact interpretation will depend on the currently held type of the interface

EDM4hep.edm4hep!TrackerHit3D — Type

Tracker hit

Author: EDM4hep authors

Fields

cellID::UInt64: ID of the sensor that created this hit

type::Int32: type of raw data hit

quality::Int32: quality bit flag of the hit

time::Float32: time of the hit

eDep::Float32: energy deposited on the hit

eDepError::Float32: error measured on eDep

position::edm4hep!Vector3d: hit position

covMatrix::edm4hep!CovMatrix3f: covariance matrix of the position (x,y,z)

EDM4hep.edm4hep!TrackerHitPlane — Type

Tracker hit plane

Author: EDM4hep authors

Fields

cellID::UInt64: ID of the sensor that created this hit

type::Int32: type of raw data hit

quality::Int32: quality bit flag of the hit

time::Float32: time of the hit

eDep::Float32: energy deposited on the hit

eDepError::Float32: error measured on eDep

u::edm4hep!Vector2f: direction of the first measurement given as (theta, phi) in spherical coordinates

v::edm4hep!Vector2f: direction of the second measurement given as (theta, phi) in spherical coordinates

du::Float32: measurement error along the direction

dv::Float32: measurement error along the direction

position::edm4hep!Vector3d: hit position

covMatrix::edm4hep!CovMatrix3f: covariance of the position (x,y,z)

EDM4hep.edm4hep!Vector2f — Type

edm4hep!Vector2f

Fields

a::Float32:

b::Float32:

EDM4hep.edm4hep!Vector3d — Type

edm4hep!Vector3d

Fields

x::Float64:

y::Float64:

z::Float64:

EDM4hep.edm4hep!Vector3f — Type

edm4hep!Vector3f

Fields

x::Float32:

y::Float32:

z::Float32:

EDM4hep.edm4hep!Vector4f — Type

Generic vector for storing classical 4D coordinates in memory. Four momentum helper functions are in edm4hep::utils

Fields

x::Float32:

y::Float32:

z::Float32:

t::Float32:

EDM4hep.edm4hep!Vertex — Type

Vertex

Author: EDM4hep authors

Fields

type::UInt32: flagword that defines the type of the vertex, see reserved bits for more information

chi2::Float32: chi-squared of the vertex fit

ndf::Int32: number of degrees of freedom of the vertex fit

position::edm4hep!Vector3f: position of the vertex

covMatrix::edm4hep!CovMatrix3f: covariance matrix of the position

algorithmType::Int32: type code for the algorithm that has been used to create the vertex

parameters::PVector{Float32}: additional parameters related to this vertex

Relations

particles::edm4hep!POD: particles that have been used to form this vertex, aka the decay particles emerging from this vertex

Methods

setParameters(object::edm4hep!Vertex, v::AbstractVector{Float32}): assign a set of values to the parameters vector member

pushToParticles(obj::edm4hep!Vertex, robj::edm4hep!POD): push related object to the particles relation

popFromParticles(obj::edm4hep!Vertex): pop last related object from particles relation

EDM4hep.RootIO.Reader — Type

The Reader structure keeps a reference to the UnROOT LazyTree and caches already built 'layouts' of the EDM4hep types. The layouts maps a set of columns in the LazyTree into an object.

EDM4hep.Histograms.H1D — Type

H1D(title::String, nbins::Int, min::Float, max::Float, unit::Symbol) Create a 1-dimensional histogram carrying the title and units.

EDM4hep.Histograms.H2D — Type

H2D(title::String, xbins::Int, xmin::Float, xmax::Float, ybins::Int, ymin::Float, ymax::Float, unit::Tuple{Symbol,Symbol}) Create a 2-dimensional histogram carrying the title and units.

EDM4hep.Histograms.H3D — Type

H3D(title::String, xbins::Int, xmin::Float, xmax::Float, ybins::Int, ymin::Float, ymax::Float, zbins::Int, zmin::Float, zmax::Float, unit::Tuple{Symbol,Symbol,Symbol}) Create a 2-dimensional histogram carrying the title and units.

Functions

EDM4hep.emptyEDStore — Method

emptyEDStore()

Empty the whole store.

EDM4hep.getEDCollection — Method

getEDCollection(::Type{ED}, collid::UInt32=0x00000000)

Get the store corresponding to the collid. If it is not specified then obtain a collid from the data type ED.

EDM4hep.hasEDCollection — Method

hasEDCollection(collid::UInt32)

Find out if the store with collid is there.

EDM4hep.initEDCollection — Method

initEDCollection(::Type{ED}) where ED

Initialize the store corresponding to type ED.

EDM4hep.RootIO.create_getter — Method

create_getter(reader::Reader, bname::String; selection=nothing)

This function creates a getter function for a given branch name. The getter function is a function that takes an event and returns a StructArray with the data of the branch. The getter function is created as a function with the name get_<branchname>. The optional parameter selection is a list of field names to be selected from the branch. If selection is not provided, all fields are selected. The user can use a list of strings, symbols or regular expressions to select the fields.

EDM4hep.RootIO.get — Method

get(reader::Reader, treename::String)

Opens a 'TTree' or 'RNTuple' in the ROOT file (typically the events tree). It returns a 'LazyTree' that allows the user to iterate over events.

EDM4hep.RootIO.get — Method

get(reader::Reader, evt::UnROOT.LazyEvent, bname::String; btype::Type=Any, register=true)

Gets an object collection by its name, with the possibility to overwrite the mapping Julia type or use the type known in the ROOT file (C++ class name). The optional key parameter register indicates if the collection needs to be registered to the EDStore.

Base.append! — Method

Base.append!(d1::AbstractAnalysisData, d2::AbstractAnalysisData)

Default function to reset the user analysis data structure in case it is not provided explicitly.

Base.empty! — Method

Base.empty!(data::AbstractAnalysisData)

Default function to reset the user analysis data structure in case it is not provided explicitly.

EDM4hep.Analysis.do_analysis! — Method

doanalysis!(data::AbstractAnalysisData, analysis, reader, events; mt::Bool=false, tasksper_thread::Int=4)

Perform an analysis on all events by executing the analysis function. The iteration will be chunked and distributed to different tasks running on different threads if the option argument mt is set to true. The results in data for all the chunks will be merged at the end of the analysis.